+++ internal/collector/datadog/metrics.go
@@ -4,3 +4,8 @@ type FixtureMetric struct {
 	Name string   `json:"name"`
 	Tags []string `json:"tags"`
 }
+
+type FixtureVolumes struct {
+	LogVolumeKnown   bool `json:"logVolumeKnown"`
+	TraceVolumeKnown bool `json:"traceVolumeKnown"`
+}